+++ lib/src/navigator/thrio_navigator.dart
@@ -176,6 +176,22 @@ abstract class ThrioNavigator {
         animated: animated,
       );
 
+  /// Pop the page in the navigation stack until the last page with `url` satisfies the `predicate`.
+  ///
+  static Future<bool> popUntil({
+    required final bool Function(String url) predicate,
+    final bool animated = true,
+  }) =>
+      ThrioNavigatorImplement.shared().popUntil(predicate: predicate, animated: animated);
+
+  /// Pop the page in the navigation stack until the first page with `url` satisfies the `predicate`.
+  ///
+  static Future<bool> popUntilFirst({
+    required final bool Function(String url) predicate,
+    final bool animated = true,
+  }) =>
+      ThrioNavigatorImplement.shared().popUntilFirst(predicate: predicate, animated: animated);
+
   /// Remove the page with `url` in the navigation stack.
   ///
   static Future<bool> remove({
@@ -186,6 +202,30 @@ abstract class ThrioNavigator {
       ThrioNavigatorImplement.shared().remove(
         url: url,
         index: index,
+        animated: animated,
+      );
+
+  /// Remove pages below the last page in the navigation stack.
+  /// Until the last page with `url` satisfies the `predicate`.
+  ///
+  static Future<bool> removeBlowUntil({
+    required final bool Function(String url) predicate,
+    final bool animated = true,
+  }) =>
+      ThrioNavigatorImplement.shared().removeBlowUntil(
+        predicate: predicate,
+        animated: animated,
+      );
+
+  /// Remove pages below the last page  in the navigation stack.
+  /// Until the first page with `url` satisfies the `predicate`.
+  ///
+  static Future<bool> removeBlowUntilFirst({
+    required final bool Function(String url) predicate,
+    final bool animated = true,
+  }) =>
+      ThrioNavigatorImplement.shared().removeBlowUntilFirst(
+        predicate: predicate,
         animated: animated,
       );