+++ CHANGELOG.md
@@ -448,3 +448,8 @@
 ## 3.7.1
 
 - feat: add popToFirst
+
+## 3.7.2
+
+- feat: add popUntil, popUntilFirst
+- feat: add removeBelowUntil, removeBelowUntilFirst